+++ app/services/provider_profile/transition_to/validated_service.rb
@@ -1,0 +1,7 @@
+class ProviderProfile < ActiveRecord::Base
+  module TransitionTo
+    class ValidatedService < GenericTransitionService
+
+    end
+  end
+end
